modify makechips.sh to be more user friendly
The makechips.sh script is supposed to download the chip information
archive from the Dediprog website. The archive's URL keeps changing as
Dediprog releases new archives and removes old ones.

In case the script is trying to access a removed archive, the Dediprog
site does not return any HTTP errors, so curl used by the script just
loads the page the Dediprog site displays and passes it to the
decompressor.

This change makes the script a bit friendlier:
  - give user a hint which package includes the 7z archiver
  - if the downloaded file does not seem to be a valid archive,
    suggest where the user can find the latest archive.
  - return nonzero exit code on 7z errors
  - allow to run it from any directory
  - make temp dir names more recognizable

TEST=tried running the script, got the following error message:
   $ ~/old_projects/em100/makechips.sh
   Downloading EM100_4.2.11.zip...
   Unpacking configs...
   No msi component found. Is http://www.dediprog.com/files/81/2968/4/EM100_4.2.11.zip a correct url?
   check http://www.dediprog.com/download?u=42&l=EM100Pro and edit /usr/local/google/home/vbendeb/old_projects/em100/makechips.sh to use the latest archive URL

Change-Id: Ib3586a493cbec688be0e03627c7ce4001642c626
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://review.coreboot.org/3945
Tested-by: build bot (Jenkins)
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>