Implement SPI trace capability for EM100 pro
This adds an experimental SPI trace command "-t".
To use it, also configure the chip type (eg. -c W25Q64CV),
and don't forget to load an image into the EM100, so
the target has something to do.
All that can be done with a single command.

The tool will run endlessly and request the EM100 pro
to report SPI traffic.

There's no interpretation yet, not even if any byte is
MISO or MOSI - the protocol doesn't seem to contain that,
so some SPI parser is required to determine at which point
the command ends and data starts.

Change-Id: I4ebc6e92cb87fb657b99fcd79d16cce4016336af
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: http://review.coreboot.org/7229
Tested-by: build bot (Jenkins)
Reviewed-by: Edward O'Callaghan <eocallaghan@alterapraxis.com>