makechips: Fix handling of SFDP/SRST data
The configuration files for each chip can contain SFDP and
SRST data, which needs to be sent to the em100 for (some of?)
these chips to work properly.

This adds the support to parse these regions and put it in the
proper form to be sent to the em100.

Unfortunately the version number for the header is the same so
the check is just done based on whether the data happens to be
present in the file.

This was tested with W25Q128FV and W25R128FV emulation which
contains both SFDP and SRST data that must be sent for the em100
to properly emulate the chip.

Change-Id: I4e6afa71bd6443a226379aae54804eb48e6551d7
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: http://review.coreboot.org/10396
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>