em100: Fix reading of data from em100
The code to read data back from the em100 needs to break the transfer
up into chunks or it will fail with 16MB parts.  This is changed to
match the code that sends the data to the em100 in chunks.

In addition the address field does not seem to be correct and in my
testing it seems to match the format that the command to send data uses.

Without this change reading back the data from a 16MB part does not work.
However I've only been able to test this with W25Q64 and W25Q128 parts
so it is possible this change is not correct for all chips either.

Change-Id: I453bf44779941584f8493d2ceba1f7deba0a4b24
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: http://review.coreboot.org/10397
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>