em100_attach: Deal with multiple devices
This patch reworks em100_attach() to deal with more than one
attached EM100pro.

According to the libusb API documentation, the function
libusb_open_device_with_vid_pid() does not require running
libusb_get_device_list(), so rework the code that in the default
case we don't do the extra work.

In this patch the behavior of the utility does not change.

Change-Id: I594db39e941f229333d6607e2a3528aad5fcb65d
Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: http://review.coreboot.org/11500
Tested-by: build bot (Jenkins)
Reviewed-by: Martin Roth <martinroth@google.com>