em100: Detect emulated flash chip
It's not clear why uploads (EM100 to host) generate a 64MiB ROM file.
It turns out that the flash chip needs to be specified at the command line, which
isn't intentional.

Detect the currently emulated SPI and use the size as defined in flashchips.

Tested on EM100Pro, will now upload ROM files matching the currently emulated target
even without specifying the flash chip on the command line.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Change-Id: Ia3fb3034a195313b8d3bb4ce24dc1779058fd08f
Reviewed-on: https://review.coreboot.org/c/em100/+/35947
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@google.com>