Make firmware update hardware version aware
Up to now, you can accidently write an EM100Pro firmware version on
an EM100Pro-G2 device. This will brick your device and can not easily
be recovered with a SPI flasher, so be careful.
Hence, the firmware update code should make sure that HW version 4 is
used to commence a firmware update.

Right now, release binaries do not contain EM100Pro-G2 firmware images,
so this version of the code bails out on anything but HW version 4.

Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Change-Id: Ia284adca4cb4d341a8d229374b6ea0b5d2a5d5c0
Reviewed-on: https://review.coreboot.org/c/em100/+/36836
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Patrick Georgi <pgeorgi@google.com>