Add network update functionality
Update chip database and firmware images over the network.
If you wish to update your chips database or firmware
archive, you can now use em100 -U to do so.

The files configs.tar.xz, firmware.tar.xz and VERSION will be
downloaded and stored in $EM100_HOME if it is set. Otherwise
$HOME/.em100 will be used as a default. If the directory does
not exist when running the command, it will be created before
downloading.

These tarballs are the same tarballs that are created with the
make tarballs target.

Change-Id: I5c378754da4ce06b63f317eb22fa237d7a616fd8
Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: https://review.coreboot.org/c/em100/+/37072
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>