makechips.sh: Analyse the Windows installer for version
Right now, the version we assume for the chips and firmware is whatever
the dediprog.com lists as the latest version (as long as the web site
formwat does not change). This is inaccurate, because you might be
operating on a very different msi archive (e.g. using a beta version,
or when operating with an older, cached binary on a builder). Instead,
use the ProductVersion string that is encoded in the binary as a trusted
source.

Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Change-Id: I9c7e48d1372bc587014e3af0a0327c70c30d7faa
Reviewed-on: https://review.coreboot.org/c/em100/+/37410
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Patrick Georgi <pgeorgi@google.com>